\name{is_rstudio_job_available}
\alias{is_rstudio_job_available}
\title{Check whether RStudio Jobs are available}
\usage{
is_rstudio_job_available()
}
\value{
Logical value indicating whether \pkg{rstudioapi} Jobs can be used.
}
\description{
Checks whether the current session can call the RStudio Jobs API.
}
